feat:[bs-48] add accounts schedulers
